Compatibility Fix for Kris
Due to concerns about Cyber's Alcohol addon becoming depricated for Hideous Destructor, the servers going down or a myriad of reasons that basically amount to him no longer working on the project, I've added it to offworld wares as it's an important part of how the "blue rum" works. I will state this multiple times that the code is not mine and I do not plan on ever claiming it is mine until maintenance or a total rewrite is needed, this inclusion is simply to reduce the amount of dependencies and potentially modify the rum's drunk effects.